product_service updated
updated the code of product_service to be consistent with  code from other services (added dto)

+++ back-end/Product_service/src/main/java/com/ishan_bhat/ProductService/entity/Product.java
@@ -20,10 +20,10 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "category_id")
-    private Integer categoryId;
+    private Long categoryId;
 
     @Column(name = "name")
     private String name;
