+++ back-end/Product_service/src/main/java/com/ishan_bhat/ProductService/repository/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    // custom query methods needed
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
 }
