+++ back-end/Product_service/src/main/java/com/ishan_bhat/ProductService/repository/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    //have to add custom query methods
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
 }
